// src/Enums/Attacks.ts
export enum PlayerAttacks {
  Validando = "Validando",
  Inyeccion = "Inyeccion",
  Attack3 = "Attack3",
  Attack4 = "Attack4",
}

export enum OponentAttacks {
  Documentacion = "Documentación incompleta",
  Version = "Regreso de version",
  Quesadilla = "Efecto quesadilla",
  RF = "RF",
  Granito = "Barra de granito",
}